ButtonHelperFatReductionInitiative - show unplayed ACs
